requirements.txt
flask
firebase-admin
flask-cors
python-dotenv
